refactor: converts telegram release to npm publish
This commit refactors the GitHub workflow. It replaces the
telegram release workflow with one that builds and publishes
to npm.

This change enables automated publishing of new versions to
the npm registry when a tagged commit is pushed to the
repository. It removes the telegram notification feature.

Also, adds new files defining parameters for serverless
functionality within the VastAI node, including endpoints and
worker groups. These parameters are exported as constants to
be used elsewhere in the node's logic. It also refactors the
api calls to use the new Vast AI api.